+++ backend/config/config.go
@@ -88,6 +88,9 @@ type Backend struct {
 	// or set to empty by the frontend if its value matches native locale
 	// as reported by the OS.
 	UserLanguage string `json:"userLanguage"`
+
+	// BtcUnit is the unit used to represents Bitcoin amounts, e.g. `BTC` or `sat`.
+	BtcUnit string `json:"btcUnit"`
 }
 
 // DeprecatedCoinActive returns the Active setting for a coin by code.  This call is should not be
@@ -217,6 +220,7 @@ func NewDefaultAppConfig() AppConfig {
 			// Copied from frontend/web/src/components/rates/rates.tsx.
 			FiatList: []string{rates.USD.String(), rates.EUR.String(), rates.CHF.String()},
 			MainFiat: rates.USD.String(),
+			BtcUnit:  "BTC",
 		},
 	}
 }